Rename `state` module to `states` to avoid collision with a state object
